// apps/legacy/src/app/lib/types.ts
'use client';
import type { Timestamp } from 'firebase/firestore';
import type { 
  EyeStyle, 
  EyeType, 
  HairStyle, 
  MouthType, 
  NoseType, 
  ShirtType, 
  GlassesType, 
  EarSize, 
  HatType,
  EyebrowType
} from 'react-nice-avatar';


// This represents the user profile document stored in Firestore
export type UserProfile = {
  id: string;
  email: string;
  firstName?: string;
  lastName?: string;
  username?: string;
  profilePicture?: string;
  totalPoints: number;
  braveCoins: number;
  questsCompleted: number;
  gender?: 'male' | 'female';
  birthday?: string;
  settings?: {
    notifications?: {
      questUpdates?: boolean;
      weeklySummary?: boolean;
    }
  }
  avatarConfig?: AvatarConfig;
  showOnLeaderboard?: boolean;
  role?: 'admin' | 'student';
};

export type AvatarConfig = {
  faceColor: string;
  hairColor: string;
  hatColor: string;
  shirtColor: string;
  bgColor: string;
  earSize: EarSize;
  eyeType: EyeType;
  eyeStyle: EyeStyle;
  eyebrowStyle: EyebrowType;
  hairStyle: HairStyle;
  hatStyle: HatType;
  mouthStyle: MouthType;
  noseStyle: NoseType;
  shirtStyle: ShirtType;
  glassesStyle: GlassesType;
};


// This represents a "deed" or "quest" submission document in Firestore
export type Deed = {
  id: string;
  userProfileId: string;
  description: string;
  photo: string;
  points: number;
  status: 'draft' | 'pending' | 'approved' | 'rejected';
  submittedAt: Timestamp;
  category: string;
};

// This is an AI-generated type, not stored in Firestore
export type VolunteerOpportunity = {
  title: string;
  description:string;
  category: string;
};

// This is based on mock-data for now
export type Reward = {
  id: string;
  name: string;
  description: string;
  cost: number;
  image: string;
};

// This represents an achievement document in Firestore
export type Achievement = {
  id: string; // This will be the document ID
  name: string;
  description: string;
  icon: string;
  unlockedAt?: string; // Should be ISO8601 string
};

// Represents a reward that a user has redeemed
export type RedeemedReward = {
  id: string; // Document ID
  rewardId: string;
  name: string;
  cost: number;
  image: string;
  redeemedAt: Timestamp;
  status: 'processing' | 'shipped' | 'delivered';
}

// Represents a notification for a user
export type Notification = {
  id: string; // Document ID
  title: string;
  description: string;
  createdAt: Timestamp;
  read: boolean;
  type: 'quest_approved' | 'achievement_unlocked';
  link?: string;
};